+++ src/data-table/columns/history.tsx
@@ -1,13 +1,11 @@
 import { IHistoryData, HistoryStatus } from "@/lib/types/data";
 import { ColumnDef } from "@tanstack/react-table";
 import { IHistoryPageState } from "@/lib/types/states";
-import { useSettings } from "@/context/settings";
-import { useTranslation } from "react-i18next";
 import { HistoryType } from "@/lib/types/enums";
-import SortableHeader from "../cells/sortable-header";
-import StatusCell from "../cells/history/status";
-import ActionsCell from "../cells/history/actions";
-import MarkAcknowledgedCell from "../cells/history/mark-acknowledged";
+import ActionsCell from "../actions/history";
+import { DetailsHeader, EventHeader, IdHeader, StatusHeader, TimestampHeader } from "../headers/history";
+import { StatusCell, MarkAcknowledgedCell, EventCell, DetailsCell } from "../cells/history";
+import { DateCell } from "../cells";
 
 export const GET_HISTORY_COLS = (
      setHistoryState: React.Dispatch<React.SetStateAction<IHistoryPageState>>,
@@ -16,54 +14,22 @@ export const GET_HISTORY_COLS = (
      const baseCols: ColumnDef<IHistoryData<"state">>[] = [
           {
                accessorKey: "timestamp",
-               header: ({column}) => {
-                    const {t} = useTranslation("table")
-                    return (
-                         <SortableHeader
-                              column={column}
-                              title={t("heading.history.timestamp")}
-                         />
-                    )
-               },
-               cell: ({getValue}) => {
-                    const {formatDate} = useSettings();
-                    return formatDate(getValue<string>())
-               }
+               header: ({column}) => <TimestampHeader column={column}/>,
+               cell: ({getValue}) => <DateCell value={getValue<string>()}/>
           },
           {
                accessorKey: "action",
-               header: ({column}) => {
-                    const {t} = useTranslation("table")
-                    return (
-                         <SortableHeader
-                              column={column}
-                              title={t("heading.history.event")}
-                         />
-                    )
-               },
-               cell: ({getValue}) => {
-                    const {t} = useTranslation("history");
-                    return t(`events.${getValue<HistoryType>()}`)
-               }
+               header: ({column}) => <EventHeader column={column}/>,
+               cell: ({getValue}) => <EventCell historyType={getValue<HistoryType>()}/>
           },
           {
                accessorKey: "details",
-               header: () => {
-                    const {t} = useTranslation("table")
-                    return t("heading.history.details")
-               },
-               cell: ({ getValue }) => (
-                    <div className="max-w-xs truncate">
-                         {getValue<string>()}
-                    </div>
-               )
+               header: () => <DetailsHeader/>,
+               cell: ({ getValue }) => <DetailsCell value={getValue<string>()}/>
           },
           {
                accessorKey: "status",
-               header: ()=>{
-                    const {t} = useTranslation("table");
-                    return t("heading.status")
-               },
+               header: () => <StatusHeader/>,
                cell: ({getValue}) => <StatusCell historyStatus={getValue<HistoryStatus>()} />
           },
           {
@@ -89,7 +55,7 @@ export const GET_HISTORY_COLS = (
           acknowledgeCol,
           {
                accessorKey: "id",
-               header: "Entry ID"
+               header: () => <IdHeader/>
           },
           ...baseCols
      ] : [