+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Improved
 - **Caching** - Cached the app name with app version into the `getAppDetails()` function
 - **The Sidebar Design** - Only icons will show if it's collapsed
+- Organized Headers and cells of data tables
 ### Fixed
 - The History Details action greying out bug
 - The bug of showing nothing on the Device Details stat after caching